🚚 Rename: 변경 파일명 PlaceHome -> PlaceHomeContents, PlaceTopNav -> PlaceHome Fixes: #5, #24
