+++ src/App.tsx
@@ -2,7 +2,7 @@ import { BrowserRouter as Router, Route, Routes } from 'react-router-dom';
 import Home from './page/Home/Home';
 import MyPage from './page/MyPage/MyPage';
 import PlaceFilter from './page/PlaceFilter/PlaceFilter';
-import PlaceTopNav from './components/BDPlace/PlaceTopNav';
+import PlaceHome from './components/BDPlace/PlaceHome';
 import Login from './page/Login/Login';
 import PrivateRoute from './components/PrivateRoute';
 import AuthProvider from './page/Login/AuthProvider';
@@ -19,7 +19,7 @@ function App() {
       <Routes>
         <Route path='/' element={<Layout />}>
           <Route index element={<Home />} />
-          <Route path='/bdPlaceHome' element={<PlaceTopNav />} />
+          <Route path='/bdPlaceHome' element={<PlaceHome />} />
           <Route path='/bdInfo' element={<Info />} />
           <Route path='/bdMag' element={<Magazine />} />
           <Route path='/mypage' element={<PrivateRoute element={MyPage} />} />
