+++ src/components/BDPlace/PlaceHome.tsx
@@ -1,5 +1,5 @@
 import { useState, useEffect, useRef } from 'react';
-import PlaceHome from '../../page/PlaceHome/PlaceHome';
+import PlaceHomeContents from '../../page/placeHome/PlaceHomeContents';
 import { twMerge } from 'tailwind-merge';
 
 const PlaceTopNav: React.FC = () => {
@@ -8,9 +8,9 @@ const PlaceTopNav: React.FC = () => {
   const indicatorRef = useRef<HTMLSpanElement | null>(null);
 
   const placeNavMenu = [
-    { name: '애개플레이스', component: <PlaceHome current={'BD'} /> },
-    { name: '펫존', component: <PlaceHome current={'pet'} /> },
-    { name: '키즈존', component: <PlaceHome current={'kid'} /> },
+    { name: '애개플레이스', component: <PlaceHomeContents current={'BD'} /> },
+    { name: '펫존', component: <PlaceHomeContents current={'pet'} /> },
+    { name: '키즈존', component: <PlaceHomeContents current={'kid'} /> },
     // ex: 컴포넌트 연결시 이렇게 사용 { name: '애개플레이스', component: <MiddleNav /> },
   ];
